Do not automatically subtract used RAM from free amount.
The amount of *currently allocated* ram (i.e. used *right now* for VMs) shouldn't be counted against the amount of free RAM when re-allocating nodes.  In effect, we are assuming that all nodes (at this stage of development) are 100% empty and ready for allocations carte blanche.

Of course, this is a complete fiction, but a useful one nonetheless.