Show "used" values for VMs, instead of the maximum allocations.
Add the --allocated option, which will draw a second set of boxes
in the output images showing the "actual" MEM/CPU usage of the VMs.
The CPU usage is a complete lie, and has a minimum of 10px in height
(otherwise, the text is unreadible.  RAM values are not fudged.